+++ src/main/java/com/hsport/wxprogram/common/util/SysuserContext.java
@@ -32,7 +32,7 @@ public class SysuserContext {
     }
 
 
-    public static Integer getUUID() {
+    public static Long getUUID() {
         String id =null;
         UUID uuid = UUID.randomUUID();
         id = uuid.toString();
@@ -41,7 +41,7 @@ public class SysuserContext {
         id = id.replace("-", "");
 
         //将随机ID换成数字
-        int num = id.hashCode();
+        Long num = (long)id.hashCode();
         //去绝对值
         num = num < 0 ? -num : num;
