+++ src/main/java/com/hsport/wxprogram/common/util/UserContext.java
@@ -29,7 +29,7 @@ public class UserContext {
         Subject currentUser = SecurityUtils.getSubject();
         return (User) currentUser.getSession().getAttribute(CURRENT_LOGIN_USER);
     }
-    public static Integer getUUID() {
+    public static Long getUUID() {
         String id =null;
         UUID uuid = UUID.randomUUID();
         id = uuid.toString();
@@ -38,7 +38,7 @@ public class UserContext {
         id = id.replace("-", "");
         System.out.println(id);
         //将随机ID换成数字
-        int num = id.hashCode();
+        Long num =(long) id.hashCode();
         //取绝对值
         num = num < 0 ? -num : num;
