+++ src/main/java/com/hsport/wxprogram/service/IJibingService.java
@@ -13,5 +13,5 @@ import org.springframework.web.bind.annotation.PathVariable;
  * @since 2019-11-26
  */
 public interface IJibingService extends IService<Jibing> {
-    public Jibing getByUserID(Integer id);
+    public Jibing getByUserID(Long id);
 }
